Add files via upload
This project solves the infamous Tower of Hanoi Puzzle. It is based off of a stack data structure that will remove the top disk from a tower and move the disk to various poles. There is both a back-end and front-end as upon running the Project Runner class, the individual movement of the disks can be seen.